<<Negocios>> Mas tipos de notificaciones

+++ GeoRed/source/Negocios/source/persistencia/Local.java
@@ -27,8 +27,8 @@ public class Local implements Serializable, Notificacion {
 	private String nombre;
 	private String descripcion;
 	
-	private float latitud;
-	private float longitud;
+	private double latitud;
+	private double longitud;
 	
 	
 	@OneToMany(mappedBy="local", cascade={CascadeType.REMOVE})
@@ -60,18 +60,18 @@ public class Local implements Serializable, Notificacion {
 	public void setDescripcion(String descripcion) {
 		this.descripcion = descripcion;
 	}   
-	public float getLatitud() {
+	public double getLatitud() {
 		return this.latitud;
 	}
 
-	public void setLatitud(float latitud) {
+	public void setLatitud(double latitud) {
 		this.latitud = latitud;
 	}   
-	public float getLongitud() {
+	public double getLongitud() {
 		return this.longitud;
 	}
 
-	public void setLongitud(float longitud) {
+	public void setLongitud(double longitud) {
 		this.longitud = longitud;
 	}
 	public Collection<Oferta> getOfertas() {
